Refactor data methods from to dashboardData.ts
